Changed code structure to add more social networks in the future
