+++ config.json
@@ -44,6 +44,8 @@
             "remove_mentions": false,
             "remove_mentions_at_char": false
         },
+        "post_frequency_minutes": 0,
+        "post_retry_timeout": 10000,
         "tags": {
             "check_for_duplicate": true,
             "limit": false,
@@ -51,8 +53,7 @@
             "remove_tags_by_length_opposite": false,
             "remove_tags_by_order": true,
             "remove_tags_by_order_opposite": false
-        },
-        "tweet_frequency_minutes": 0
+        }
     },
     "steem_accounts": {
         "posts": [
@@ -67,11 +68,13 @@
         "https://rpc.steemviz.com",
         "https://steemd.minnowsupportproject.org"
     ],
+    "targets": {
+        "twitter": true
+    },
     "template": {
         "post": "{title}::1 {tags}::2 {link}",
         "resteem": "{title}::1 [by {author}]::3 {tags}::2 {link}",
         "tweet_like": "{title} {tags}::1"
     },
-    "tweet_retry_timeout": 10000,
     "twitter_handle": "RagePeanut_"
 }
