+++ utils/parser.js
@@ -8,7 +8,7 @@ const modifiers = {
 
 /**
  * Parses a template into a structure
- * @param {string} template The template of the tweet
+ * @param {string} template The template of the message
  * @param {object} values The variables' values
  * @returns {object} The parsed structure
  */
@@ -51,6 +51,58 @@ function parse(template, values) {
 }
 
 /**
+ * Removes or modifies the least important element from a structure
+ * @param {object} structure The structure of the part of a message
+ * @returns {object} The passed structure with one modified element
+ */
+function removeLeastImportant(structure) {
+    if(structure.children.length > 0) {
+        const leastImportantChild = structure.children[structure.children.length - 1];
+        const originalLeastImportantChildParsed = leastImportantChild.parsed;
+        if(leastImportantChild.children.length > 0) {
+            leastImportantChild = removeLeastImportant(leastImportantChild);
+            structure.parsed = structure.parsed.replace(originalLeastImportantChildParsed, leastImportantChild.parsed);
+        } else {
+            switch(leastImportantChild.variable) {
+                case 'tags':
+                    // If set to true, removes tags by order of importance (last tag removed first)
+                    if(settings.tags.remove_tags_by_order) leastImportantChild.parsed = leastImportantChild.parsed.replace(/ ?#[A-Za-z\d-]+$/, '');
+                    // If set to true, removes tags by order of importance (first tag removed first)
+                    else if(settings.tags.remove_tags_by_order_opposite) leastImportantChild.parsed = leastImportantChild.parsed.replace(/^#[A-Za-z\d-]+ ?/, '');
+                    // If set to true, removes tags by length (smallest removed first)
+                    else if(settings.tags.remove_tags_by_length) {
+                        const tags = leastImportantChild.parsed.split(' ');
+                        const smallestTag = tags.reduce((a, b) => a.length < b.length ? a : b);
+                        leastImportantChild.parsed = tags.filter(tag => tag !== smallestTag).join(' ');
+                    // If set to true, removes tags by length (longest removed first)
+                    } else if(settings.tags.remove_tags_by_length_opposite) {
+                        const tags = leastImportantChild.parsed.split(' ');
+                        const longestTag = tags.reduce((a, b) => a.length > b.length ? a : b);
+                        leastImportantChild.parsed = tags.filter(tag => tag !== longestTag).join(' ');
+                    } else leastImportantChild.parsed = '';
+                    break;
+                case 'title':
+                    leastImportantChild.parsed = leastImportantChild.parsed.replace(/.(.{3})$/, (match, lastChars) => {
+                        if(lastChars === '...') return '...';
+                        else return match[0] + '...';
+                    });
+                    if(leastImportantChild.parsed === '...') leastImportantChild.parsed = '';
+                    break;
+                default:
+                    leastImportantChild.parsed = '';
+                    break;
+            }
+            structure.parsed = structure.parsed.replace(originalLeastImportantChildParsed, leastImportantChild.parsed).replace(/ +/, ' ');
+            if(leastImportantChild.parsed === '') {
+                structure.raw = structure.raw.replace(leastImportantChild.raw, '').replace(/ +/, ' ');
+                structure.children.pop();
+            }
+        }
+    }
+    return structure;
+}
+
+/**
  * Processes a variable based on its value and modifier
  * @param {string} variable The variable to process
  * @param {string} modifier The modifier's name
@@ -88,5 +140,6 @@ function processVariable(variable, modifier, values) {
 }
 
 module.exports = {
-    parse: parse
+    parse: parse,
+    removeLeastImportant
 }
